[primitives] rename key2 -> key and key_ptr2 -> key_ptr module
